refactor: Replace and simplify security configuration
- Renamed `SpringSecurityConfig` class to `SecurityConfig`.
- Updated security configuration to use `UserDetailsLoaderService` and
`JwtService`.
- Refactored `SecurityFilterChain` to include specific CSRF exclusions
for `/api/auth/register` and `/api/auth/login`.
- Enhanced logging with `Logger` to provide better visibility into
security configuration steps.
